Real LLM providers behind the same protocol
Step 11 of the Phase 1 build order.

- Anthropic, via the official SDK. Uses output_config.format to constrain the
  reply to the rerank JSON schema server-side, and sets no temperature --
  current Claude models reject the parameter, and determinism is not something
  to promise from a real model anyway; that guarantee belongs to the fake.
  A max_tokens truncation is reported as its own error rather than fed to the
  parser, which would spend the one repair retry on a doomed parse.
- OpenAI-compatible /chat/completions, requesting
  response_format=json_object -- the one JSON-constraining feature these
  endpoints implement consistently. It guarantees syntax, not the schema.

Server-side constraint does not replace the strict parser, the repair retry
or the hallucinated-code guard. "Unlikely to be malformed" is not a property
an audit trail should rest on, and no schema can constrain *which* codes come
back, which is what the guard is about.

One renderer builds the model input for every provider, so "what the model
saw" means the same thing regardless of which one produced a proposal.
Synonym lists are truncated per candidate so one richly-described concept
cannot crowd out the others.

Live smoke tests run only when a key is present and skip otherwise, so a
clone with no credentials still gets a green suite; provider structure and
error handling are covered offline with a stubbed transport.

Also: ruff and mypy --strict now pass clean across app/.

+++ app/pipeline/map_term.py
@@ -140,9 +140,7 @@ def map_term(
     latency_ms_rerank = _elapsed_ms(rerank_started)
 
     top = result.top if result is not None else None
-    suggested_code = None if result is None or result.no_good_match else (
-        top.code if top else None
-    )
+    suggested_code = None if result is None or result.no_good_match else (top.code if top else None)
     model_confidence = None if suggested_code is None or top is None else top.confidence
 
     # ------------------------------------------------------------- audit row
@@ -187,4 +185,4 @@ def _require_loaded(session: Session, system: str, version: str) -> None:
 
 
 def _elapsed_ms(started: float) -> int:
-    return int(round((time.perf_counter() - started) * 1000))
+    return round((time.perf_counter() - started) * 1000)
